refactor: extract token-related auth logic into token service
