+++ src/services/auth/index.ts
@@ -1,2 +1,1 @@
 export { createAuthService } from './service';
-export type { WithRetry } from './service';
